+++ EasyReflectometryApp/Gui/Components/AnalysisDataChartBokeh.qml
@@ -1,11 +1,12 @@
 import QtQuick 2.13
 
-import easyAppGui.Style 1.0 as EaStyle
-import easyAppGui.Charts 1.0 as EaCharts
+import easyApp.Gui.Style 1.0 as EaStyle
+import easyApp.Gui.Charts 1.0 as EaCharts
 
+import Gui.Components 1.0 as ExComponents
 import Gui.Globals 1.0 as ExGlobals
 
-EaCharts.BaseBokeh {
+ExComponents.BaseBokeh {
     measuredData: ExGlobals.Constants.proxy.plotting1d.bokehMeasuredDataObj
     calculatedData: ExGlobals.Constants.proxy.plotting1d.bokehCalculatedDataObj
     sldData: ExGlobals.Constants.proxy.plotting1d.bokehAnalysisSldDataObj
@@ -14,11 +15,7 @@ EaCharts.BaseBokeh {
     sldPlotRanges: ExGlobals.Constants.proxy.plotting1d.analysisSldPlotRangesObj
 
     xMainAxisTitle: "q (Å⁻¹)"
-    yMainAxisTitle: {
-        let title = 'R(q)calc'
-        if (hasMeasuredData) title = 'R(q)meas, R(q)calc'
-        return title
-    }
+    yMainAxisTitle: ExGlobals.Constants.proxy.simulation.yMainAxisTitle
     xSldAxisTitle: "z (Å)"
     ySldAxisTitle: "SLD (10⁻⁶Å⁻²)"