+++ EasyReflectometryApp/Gui/Components/ExperimentDataChartBokeh.qml
@@ -1,15 +1,16 @@
 import QtQuick 2.13
 
-import easyAppGui.Charts 1.0 as EaCharts
+import easyApp.Gui.Charts 1.0 as EaCharts
 
+import Gui.Components 1.0 as ExComponents
 import Gui.Globals 1.0 as ExGlobals
 
-EaCharts.BaseBokeh {
+ExComponents.BaseBokeh {
     measuredData: ExGlobals.Constants.proxy.plotting1d.bokehMeasuredDataObj
 
     plotRanges: ExGlobals.Constants.proxy.plotting1d.experimentPlotRangesObj
     sldPlotRanges: ExGlobals.Constants.proxy.plotting1d.sldPlotRangesObj
 
     xMainAxisTitle: "q (Å⁻¹)"
-    yMainAxisTitle: "R(q)"
+    yMainAxisTitle: ExGlobals.Constants.proxy.simulation.yMainAxisTitle
 }