+++ EasyReflectometryApp/Gui/Components/SampleChartBokeh.qml
@@ -1,11 +1,11 @@
 import QtQuick 2.13
 
-import easyAppGui.Style 1.0 as EaStyle
-import easyAppGui.Charts 1.0 as EaCharts
+import easyApp.Gui.Style 1.0 as EaStyle
 
+import Gui.Components 1.0 as ExComponents
 import Gui.Globals 1.0 as ExGlobals
 
-EaCharts.BaseBokeh {
+ExComponents.BaseBokeh {
     calculatedData: ExGlobals.Constants.proxy.plotting1d.bokehPureDataObj
     sldData: ExGlobals.Constants.proxy.plotting1d.bokehSampleSldDataObj
 
@@ -13,11 +13,7 @@ EaCharts.BaseBokeh {
     sldPlotRanges: ExGlobals.Constants.proxy.plotting1d.sampleSldPlotRangesObj
 
     xMainAxisTitle: "q (Å⁻¹)"
-    yMainAxisTitle: {
-        let title = 'R(q)calc'
-        if (hasMeasuredData) title = 'R(q)meas, R(q)calc'
-        return title
-    }
+    yMainAxisTitle: ExGlobals.Constants.proxy.simulation.yMainAxisTitle
     xSldAxisTitle: "z (Å)"
     ySldAxisTitle: "SLD (10⁻⁶Å⁻²)"