+++ EasyReflectometryApp/Gui/Pages/Sample/SurfactantTable.qml
@@ -2,8 +2,8 @@ import QtQuick 2.14
 import QtQuick.Controls 2.14
 import QtQuick.XmlListModel 2.13
 
-import easyAppGui.Components 1.0 as EaComponents
-import easyAppGui.Style 1.0 as EaStyle
+import easyApp.Gui.Components 1.0 as EaComponents
+import easyApp.Gui.Style 1.0 as EaStyle
 
 import Gui.Globals 1.0 as ExGlobals