+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Enabling the docs and bug reporting buttons [7d29276](https://github.com/easyScience/EasyReflectometryApp/pull/79/commits/7d292761e96860b5b8ce74365174e38f79bbaf80)
 - Fix the reset state button [a1db4ac](https://github.com/easyScience/EasyReflectometryApp/pull/79/commits/a1db4ac10bbd5f90c764f28336cbfecb26671a03)
 - Parameter quick filters [27a64b6](https://github.com/easyScience/EasyReflectometryApp/pull/79/commits/27a64b615d016713478817119e45133b4af50695)
+- Start using easyApp []()
+- Add plotting in R(q)q^4 []()
 
 ### Changes 0.0.4